chore(runtime): split date.rs under the 2000-line CI cap
date.rs grew to 2008 lines (pushed over by the Temporal tail in #4923), tripping
scripts/check_file_size.sh and failing the required lint gate on every open PR
(the check runs against each PR's merge with current main).

Extract the date-string parsing cluster - parse_date_string + its per-grammar
helpers (parse_iso8601, parse_rfc_or_named, parse_tz_offset, normalize_millis,
month_from_name, FULL_MONTHS) - into a new crates/perry-runtime/src/date/parse.rs
submodule (Rust 2018 keeps date.rs alongside date/parse.rs, no rename). date.rs
declares `mod parse;` and re-exports the single entry point via
`use parse::parse_date_string;`; the per-grammar helpers stay private to the
submodule. Shared time math (make_utc_ms, time_clip,
timestamp_to_local_components) stays in the parent and is reached via super::
(a child module can see its ancestor's private items, so no visibility was
widened beyond parse_date_string).

Pure code movement, no behavior change. date.rs 2008 -> 1675 lines,
date/parse.rs 347. fmt clean; all date tests pass except the pre-existing
test_full_year_setters_revive_invalid_date_only, which is timezone-dependent
(asserts local hours == 0 after a local setFullYear) and fails identically on
clean main on a non-UTC host - CI runs UTC and passes it.
