+++ Cargo.toml
@@ -201,7 +201,7 @@ codegen-units = 16
 opt-level = "s"       # Optimize for size in stdlib
 
 [workspace.package]
-version = "0.5.1154"
+version = "0.5.1155"
 edition = "2021"
 license = "MIT"
 repository = "https://github.com/PerryTS/perry"
